포스트 상세 정보 받는 api 작성, valid id 검사 로직 모두 추가.

+++ src/main/java/capstone/placer/post/PostDetailMapper.java
@@ -1,7 +1,7 @@
 package capstone.placer.post;
 
 public interface PostDetailMapper {
-    PostDetail getDetail(int postId);
+    PostDetail getDetail(long postId);
 
     PostDetail insert(PostDetail postDetail);
 }
